Move table tests to separate file
